Getting spongebob up and going
